+++ src/views/index.ts
@@ -1,5 +1,4 @@
-export * from './AddItem';
 export * from './Home';
 export * from './Layout';
 export * from './List';
-export * from './Summary';
+export * from './Export';
